+++ R/varCompTest.lme.R
@@ -15,14 +15,13 @@ varCompTest.lme <- function(m1,m0,control = list(M=5000,parallel=T,nb_cores=1,B=
   
   message("Variance components testing in mixed effects models")
   
-  pkg <- pckName(m1)
   randm0 <- !(max(class(m0) %in% c("lm","glm","nls"))) # are there any random effect under H0?
   
   # Extract data structure
   msdata <- extractStruct(m1,m0,randm0)
   
   if (length(msdata$nameVarTested)==1){
-    message(paste("Testing that the variance of",msdata$nameVarTested,"is null\n"))
+    message(paste("Testing that the variance of",msdata$nameVarTested,"is null"))
   }else if (length(msdata$nameVarTested) > 1){
     message(paste("Testing that the variances of",paste(msdata$nameVarTested,sep="",collapse = " and "),"are null\n"))
   }else{
@@ -99,25 +98,6 @@ varCompTest.lme <- function(m1,m0,control = list(M=5000,parallel=T,nb_cores=1,B=
   uppboundpval <- (1/2)*sum(stats::pchisq(lrt,cbs.df.dims$df[(length(cbs.df.dims$df)-1):length(cbs.df.dims$df)],lower.tail = F))
   lowboundpval <- (1/2)*sum(stats::pchisq(lrt,cbs.df.dims$df[1:2],lower.tail = F))
   
-  # Print results
-  message(paste("Likelihood ratio test statistic: \n LRT = ",format(lrt,digits=5),
-                "\n\nLimiting distribution:"))
-  if (length(cbs.df.dims$df) > 1){
-    message(paste("mixture of",length(cbs.df.dims$df),"chi-bar-square distributions with degrees of freedom",paste(cbs.df.dims$df,sep="",collapse = ", "),"\n"))
-    
-    if (pval.comp %in% c("approx","both")){
-      message(paste(" associated weights and sd: ",paste(paste(round(cbs.weights.sample$weights,3)," (",round(cbs.weights.sample$sdWeights,3),")",sep=""),sep="",collapse = ", "),
-                    "\n\np-value (from estimated weights):",format(pvalue1,digits = 5)))
-      
-      if (length(cbs.weights.sample$randomCBS) > 0) message(paste("\np-value (from simulated chi-bar-square distribution):",format(pvalue2,digits=5),"\n"))
-    }
-    
-    if (pval.comp %in% c("bounds","both")) message(paste("lower-bound for p-value:",format(lowboundpval,digits=5)," upper bound for p-value:",format(uppboundpval,digits=5)))
-  }else{
-    message(paste0("  chi-bar-square distributions with ",cbs.df.dims$df," degree of freedom\n"))
-    message(paste0("  p-value: ",format(pvalue1,digits=5)))
-  }
-  
   # create results, object of class htest
   null.value <- rep(0,length(msdata$nameVarTested)+length(msdata$nameFixedTested))
   if (length(msdata$nameFixedTested)==0){
@@ -131,10 +111,10 @@ varCompTest.lme <- function(m1,m0,control = list(M=5000,parallel=T,nb_cores=1,B=
   results <- list(statistic=c(LRT=lrt),
                   null.value=null.value,
                   alternative=alternative,
-                  parameters=list(df=cbs.df.dims$df,weights=cbs.weights.sample$weights,FIM=fim),
+                  parameters=list(df=cbs.df.dims$df,weights=cbs.weights.sample$weights,sdweights=cbs.weights.sample$sdWeights,FIM=fim),
                   method="Likelihood ratio test for variance components in mixed effects models",
                   p.value=c(pvalue.weights=pvalue1,pvalue.sample=pvalue2,pvalue.lowerbound=lowboundpval,pvalue.upperbound=uppboundpval))
-  class(results) <- "htest"
+  class(results) <- c("vctest","htest")
   
   invisible(results)
 }